test(ctm): mark sharding unit tests core, parity slow (split into own file)
